feat: First functional game UI
The UI rules around taking turns, deciding who wins, who should go
next, etc. are in now. It's using the APIs to update the training
data and get predictions.

The code and UI are both ugly as sin, so probably needs a tidy up.
But the basics are almost there.