feat: Displaying training data during the game
Initial placeholder for what we could show during the game. The
objective is to give the player a clear way of seeing how the
bot makes better choices the more training it's had.

Ideally, we need some way of showing how it's performance is
improving over time.

For now, I've grabbed what might be two useful data points:
- How much training this bot has had
- How many hands has the computer played and won in a row (with that
resetting when the computer loses a turn)

Contributes to #3

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>