feat: Add a side column with game info
Instead of displaying the outcome of each hand in a dialog that
gets in the way, it moves this info to a side column.

It also adds a place for us to put some running statistics.

I'm thinking a tiny version of the previous hand could be useful
but haven't updated the tt-card directive to support a thumbnail
view so I'm just using static placeholder images.

Closes #4