feat: Add place for educational text in the sidebar
I've set it to not change on every turn, as that will be distracting.
It should change messages every 4 turns, cycling round a set of
messages in the code.

I still need to write the messages, but I've put some placeholders
in to explain the sorts of things we could use it for.